+++ Wangscape/tilegen/partition/TilePartitionerGradient.cpp
@@ -25,7 +25,7 @@ void TilePartitionerGradient::makePartition(TilePartition & regions,
                    mOptions.tileFormat.resolution);
         masks.push_back(img);
     }
-    std::vector<float> weights(corners.size(), 0.f);
+    std::vector<double> weights(corners.size(), 0.);
     const int resolution_sub_1 = mOptions.tileFormat.resolution - 1;
     const int quarter_res = mOptions.tileFormat.resolution / 4;
     alpha::AlphaCalculatorLinear ac;
@@ -33,10 +33,10 @@ void TilePartitionerGradient::makePartition(TilePartition & regions,
     {
         for (size_t y = 0; y < mOptions.tileFormat.resolution; y++)
         {
-            weights[0] = (float)gradientWeight(x, y, 0, 0, quarter_res);
-            weights[1] = (float)gradientWeight(x, y, 0, resolution_sub_1, quarter_res);
-            weights[2] = (float)gradientWeight(x, y, resolution_sub_1, 0, quarter_res);
-            weights[3] = (float)gradientWeight(x, y, resolution_sub_1, resolution_sub_1, quarter_res);
+            weights[0] = (double)gradientWeight(x, y, 0, 0, quarter_res);
+            weights[1] = (double)gradientWeight(x, y, 0, resolution_sub_1, quarter_res);
+            weights[2] = (double)gradientWeight(x, y, resolution_sub_1, 0, quarter_res);
+            weights[3] = (double)gradientWeight(x, y, resolution_sub_1, resolution_sub_1, quarter_res);
             ac.updateAlphas(weights);
             const auto& alphas = ac.getAlphas();
             for (size_t i = 0; i < masks.size(); i++)
